Add ability to publish package without a pubspec. Review URL: https://codereview.chromium.org//11280238
git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@15544 260f80e4-7a28-3924-810f-c04153c831b5

+++ tools/publish_pkg.py
@@ -7,6 +7,8 @@
 # Script to push a package to pub. 
 #
 # Usage: publish_pkg.py pkg_dir
+#
+# "pub" must be in PATH.
 
 
 import os
@@ -45,6 +47,8 @@ def Main(argv):
   build = ReadVersion(versionFile, 'BUILD')
   patch = ReadVersion(versionFile, 'PATCH')
   
+  # bleeding_edge has a fixed version number of 0.1.x.y . Don't allow users
+  # to publish packages from bleeding_edge.
   if major == 0 and minor <= 1:
     print 'Error: Do not run this script from a bleeding_edge checkout.'
     return -1
@@ -53,25 +57,63 @@ def Main(argv):
 
   tmpDir = tempfile.mkdtemp()
   pkgName = argv[1].split('/').pop()
-  shutil.copytree(os.path.join(HOME, argv[1]), 
-                  os.path.join(tmpDir, pkgName))
 
-  # Add version to pubspec file.
   pubspec = os.path.join(tmpDir, pkgName, 'pubspec.yaml')
-  pubspecFile = open(pubspec)
-  lines = pubspecFile.readlines()
-  pubspecFile.close()
-  pubspecFile = open(pubspec, 'w')
-  foundVersion = False
-  for line in lines:
-    if line.startswith('version:'):
-      foundVersion = True
-    if line.startswith('description:') and not foundVersion:
+
+  if os.path.exists(os.path.join(HOME, argv[1], 'pubspec.yaml')):
+    #
+    # If pubspec.yaml exists, add the SDK's version number if
+    # no version number is present.
+    #
+    shutil.copytree(os.path.join(HOME, argv[1]), 
+                    os.path.join(tmpDir, pkgName))
+    with open(pubspec) as pubspecFile:
+      lines = pubspecFile.readlines()
+    with open(pubspec, 'w') as pubspecFile:
+      foundVersion = False
+      inDependencies = False
+      for line in lines:
+        if line.startswith('dependencies:'):
+          inDependencies = True
+        elif line[0].isalpha():
+          inDependencies = False
+        if line.startswith('version:'):
+          foundVersion = True
+        if inDependencies:
+          #
+          # Within dependencies, don't print line that start with "    sdk:"
+          # and strip out "{ sdk: package_name }".
+          #
+          if not line.startswith('    sdk:'):
+            line = re.sub(r'{(\s*)sdk:(\s+)([a-z0-9_]+)(\s*)}', '', line)
+            pubspecFile.write(line)
+        else:
+          pubspecFile.write(line)
+      if not foundVersion:
+        pubspecFile.write('\nversion: ' + version + '\n')
+  else:
+    #
+    # If there's a lib/ directory in the package, copy the package.
+    # Otherwise, move the package's contents to lib/.
+    #
+    if os.path.exists(os.path.join(HOME, argv[1], 'lib')):
+      shutil.copytree(os.path.join(HOME, argv[1]), 
+                      os.path.join(tmpDir, pkgName))
+    else:
+      os.makedirs(os.path.join(tmpDir, pkgName))
+      shutil.copytree(os.path.join(HOME, argv[1]),
+                      os.path.join(tmpDir, pkgName, 'lib'))
+
+    # Create pubspec.yaml .
+    with open(pubspec, 'w') as pubspecFile:
+      pubspecFile.write('name: ' + pkgName + '_unsupported\n')
       pubspecFile.write('version: ' + version + '\n')
-    if not line.startswith('    sdk:'):
-      pubspecFile.write(line)
-  pubspecFile.close()
-      
+      pubspecFile.write("description: >\n")
+      pubspecFile.write('  A completely unsupported clone of Dart SDK library\n')
+      pubspecFile.write('  ' + argv[1] + ' . This package will change in\n')
+      pubspecFile.write('  unpredictable/incompatible ways without warning.\n')
+      pubspecFile.write('dependencies:\n')
+
   # Replace '../*/pkg' imports and parts.
   for root, dirs, files in os.walk(os.path.join(tmpDir, pkgName)):
     for name in files:
@@ -79,8 +121,7 @@ def Main(argv):
         ReplaceInFiles([os.path.join(root, name)], 
             [(r'(import|part)(\s+)(\'|")(\.\./)+pkg/', r'\1\2\3package:')])
   
-  print 'publishing version ' + version + ' of ' + argv[1] + ' to pub\n'
-  print tmpDir
+  print 'publishing version ' + version + ' of ' + argv[1] + ' to pub.\n'
   subprocess.call(['pub', 'publish'], cwd=os.path.join(tmpDir, pkgName))
   shutil.rmtree(tmpDir)
